Dev fixing redis configuration to run on 3 computers (#421)
* Fixing Redis configuration to run on 3 computers

* fiexes redis and evac

* remove useless

* redis fixes

+++ installer/sql.sql
@@ -38,7 +38,7 @@ CREATE TABLE simulations ( ---{{{
     iteration smallint,
     host text,
     job_id text,
-    run_time smallint,
+    run_time int,
     fireorig text,
     fireorigname text,
     heat_detectors text,
